+++ CHANGES.rst
@@ -9,6 +9,9 @@ Changelog
 
         DNS: dnspython not found. Can not use SRV lookup.
 
+-   Pin some recently updated dependencies of Celery that caused unit test
+    failures: amqp <= 2.4.2, kombu <= 4.5.0, vine <= 1.3.0.
+
 0.6.1 (2019-05-24)
 ------------------